Move getRealPath() to FileSystemPathUtils.
