+++ src/client/common/platform/fs-paths.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+import * as fs from 'fs-extra';
 import * as nodepath from 'path';
 import { getOSType, OSType } from '../utils/platform';
 import {
@@ -122,6 +123,15 @@ export class FileSystemPathUtils {
         path1 = this.paths.normCase(path1);
         path2 = this.paths.normCase(path2);
         return path1 === path2;
+    }
+
+    public async getRealPath(filename: string): Promise<string> {
+        try {
+            return await fs.realpath(filename);
+        } catch {
+            // We ignore the error.
+            return filename;
+        }
     }
 
     public getDisplayName(pathValue: string, cwd?: string): string {
